Use value type of ONNX tensor arguments
