+++ model-integration/src/main/java/ai/vespa/rankingexpression/importer/tensorflow/TensorConverter.java
@@ -28,7 +28,7 @@ public class TensorConverter {
     }
 
     private static Tensor toVespaTensor(org.tensorflow.Tensor<?> tfTensor, String dimensionPrefix) {
-        TensorType type = toVespaTensorType(tfTensor, dimensionPrefix);
+        TensorType type = TypeConverter.typeFrom(tfTensor, dimensionPrefix);
         Values values = readValuesOf(tfTensor);
         IndexedTensor.BoundBuilder builder = (IndexedTensor.BoundBuilder)Tensor.Builder.of(type);
         for (int i = 0; i < values.size(); i++)
@@ -52,16 +52,6 @@ public class TensorConverter {
             builder.cellByDirectIndex(i, values.get(i));
         }
         return builder.build();
-    }
-
-    private static TensorType toVespaTensorType(org.tensorflow.Tensor<?> tfTensor, String dimensionPrefix) {
-        TensorType.Builder b = new TensorType.Builder(toValueType(tfTensor.dataType()));
-        int dimensionIndex = 0;
-        for (long dimensionSize : tfTensor.shape()) {
-            if (dimensionSize == 0) dimensionSize = 1; // TensorFlow ...
-            b.indexed(dimensionPrefix + (dimensionIndex++), dimensionSize);
-        }
-        return b.build();
     }
 
     public static Long tensorSize(TensorType type) {
@@ -106,21 +96,6 @@ public class TensorConverter {
                 return new ProtoDoubleValues(tensorProto);
         }
         throw new IllegalArgumentException("Unsupported data type in attribute tensor import");
-    }
-
-    /** TensorFlow has two different DataType classes. This must be kept in sync with TypeConverter.toValueType */
-    static TensorType.Value toValueType(DataType dataType) {
-        switch (dataType) {
-            case FLOAT: return TensorType.Value.FLOAT;
-            case DOUBLE: return TensorType.Value.DOUBLE;
-            // Imperfect conversion, for now:
-            case BOOL: return TensorType.Value.FLOAT;
-            case INT32: return TensorType.Value.FLOAT;
-            case UINT8: return TensorType.Value.FLOAT;
-            case INT64: return TensorType.Value.DOUBLE;
-            default: throw new IllegalArgumentException("A TensorFlow tensor with data type " + dataType +
-                                                        " cannot be converted to a Vespa tensor type");
-        }
     }
 
     /** Allows reading values from buffers of various numeric types as bytes */
