+++ model-integration/src/main/java/ai/vespa/rankingexpression/importer/tensorflow/TypeConverter.java
@@ -9,8 +9,6 @@ import org.tensorflow.framework.DataType;
 import org.tensorflow.framework.NodeDef;
 import org.tensorflow.framework.TensorShapeProto;
 
-import java.util.List;
-
 /**
  * Converts and verifies TensorFlow tensor types into Vespa tensor types.
  *
@@ -37,6 +35,32 @@ class TypeConverter {
         }
     }
 
+    static OrderedTensorType typeFrom(NodeDef node) {
+        String dimensionPrefix = "d"; // standard naming convention: d0, d1, ...
+        TensorShapeProto shape = tensorFlowShape(node);
+        OrderedTensorType.Builder builder = new OrderedTensorType.Builder(toValueType(tensorFlowValueType(node)));
+        for (int i = 0; i < shape.getDimCount(); ++ i) {
+            String dimensionName = dimensionPrefix + i;
+            TensorShapeProto.Dim tensorFlowDimension = shape.getDim(i);
+            if (tensorFlowDimension.getSize() >= 0) {
+                builder.add(TensorType.Dimension.indexed(dimensionName, tensorFlowDimension.getSize()));
+            } else {
+                builder.add(TensorType.Dimension.indexed(dimensionName));
+            }
+        }
+        return builder.build();
+    }
+
+    static TensorType typeFrom(org.tensorflow.Tensor<?> tfTensor, String dimensionPrefix) {
+        TensorType.Builder b = new TensorType.Builder(toValueType(tfTensor.dataType()));
+        int dimensionIndex = 0;
+        for (long dimensionSize : tfTensor.shape()) {
+            if (dimensionSize == 0) dimensionSize = 1; // TensorFlow ...
+            b.indexed(dimensionPrefix + (dimensionIndex++), dimensionSize);
+        }
+        return b.build();
+    }
+
     private static TensorShapeProto tensorFlowShape(NodeDef node) {
         AttrValue attrValueList = node.getAttrMap().get("_output_shapes");
         if (attrValueList == null)
@@ -59,27 +83,7 @@ class TypeConverter {
         return attrValueList.getList().getType(0); // support multiple outputs?
     }
 
-    static OrderedTensorType fromTensorFlowType(NodeDef node) {
-        return fromTensorFlowType(node, "d");  // standard naming convention: d0, d1, ...
-    }
-
-    private static OrderedTensorType fromTensorFlowType(NodeDef node, String dimensionPrefix) {
-        TensorShapeProto shape = tensorFlowShape(node);
-        OrderedTensorType.Builder builder = new OrderedTensorType.Builder(toValueType(tensorFlowValueType(node)));
-        for (int i = 0; i < shape.getDimCount(); ++ i) {
-            String dimensionName = dimensionPrefix + i;
-            TensorShapeProto.Dim tensorFlowDimension = shape.getDim(i);
-            if (tensorFlowDimension.getSize() >= 0) {
-                builder.add(TensorType.Dimension.indexed(dimensionName, tensorFlowDimension.getSize()));
-            } else {
-                builder.add(TensorType.Dimension.indexed(dimensionName));
-            }
-        }
-        return builder.build();
-    }
-
-    /** TensorFlow has two different DataType classes. This must be kept in sync with TensorConverter.toValueType */
-    static TensorType.Value toValueType(DataType dataType) {
+    private static TensorType.Value toValueType(DataType dataType) {
         switch (dataType) {
             case DT_FLOAT: return TensorType.Value.FLOAT;
             case DT_DOUBLE: return TensorType.Value.DOUBLE;
@@ -95,6 +99,20 @@ class TypeConverter {
             case DT_UINT16: return TensorType.Value.FLOAT;
             case DT_UINT32: return TensorType.Value.FLOAT;
             case DT_UINT64: return TensorType.Value.DOUBLE;
+            default: throw new IllegalArgumentException("A TensorFlow tensor with data type " + dataType +
+                                                        " cannot be converted to a Vespa tensor type");
+        }
+    }
+
+    private static TensorType.Value toValueType(org.tensorflow.DataType dataType) {
+        switch (dataType) {
+            case FLOAT: return TensorType.Value.FLOAT;
+            case DOUBLE: return TensorType.Value.DOUBLE;
+            // Imperfect conversion, for now:
+            case BOOL: return TensorType.Value.FLOAT;
+            case INT32: return TensorType.Value.FLOAT;
+            case UINT8: return TensorType.Value.FLOAT;
+            case INT64: return TensorType.Value.DOUBLE;
             default: throw new IllegalArgumentException("A TensorFlow tensor with data type " + dataType +
                                                         " cannot be converted to a Vespa tensor type");
         }
